[animation] Move the alpha value to floating point
The current Alpha value is an unsigned integer that can be used
implicitly as a fixed point value. This makes writing an alpha
function overshooting below and above the current range basically
impossible without complicating an already complex code, and
creating weird corner cases.

For this reason, the Alpha value should be defined as a floating
point normalized value, spanning a range between 0.0 and 1.0; in
order to allow overshooting, the valid range is extended one unit
below and one unit above, thus making it -1.0 .. 2.0.

This commit updates the various users of the ClutterAlpha API
and the tests cases.

This commit also removes all the current alpha functions exposed
in the public API.

+++ clutter/clutter-behaviour.h
@@ -113,7 +113,7 @@ struct _ClutterBehaviourClass
   /*< public >*/
   /* vfunc, not signal */
   void (*alpha_notify) (ClutterBehaviour *behave,
-                        guint32           alpha_value);
+                        gdouble           alpha_value);
 
   /* signals */
   void (*applied)  (ClutterBehaviour *behave,
